Move a few tests into the general tokenizer
